spi: spidev: Add spidev patch to silence warning about spidev listing in DT
spidev is the userspace interface to the SPI subsystem.
many other platforms, use a compatible string of "spidev"
to cause the spidev module to be loaded. A new commit in
the upstream 4.1 branch (956b200) demonstrates the author's
objection to this practice by causing the module to WARN
when it detects that usage. Add the "spidev" string to compatible
to silence this warning.

Change-Id: I3c904bb5520c794c3431093b9bffe7c9293f8604
Signed-off-by: Lin Huang <hl@rock-chips.com>
(cherry picked from commit da404b38fa61eb09b1a5e7c2de59fbaccdafd57b)